Remove all dependencies
By removing enumerable-statistics we break two things:

- Histogram support
- Boxplot support

This buys us:

- Not relying on monkeypatching (this decreases invasiveness of gem and makes the seams more explicit)
- Not artificially restricting target Ruby environments to only Ruby 2.4+

This could possibly ship as is and then be re-bundled into a `unicode_plot` with just the addition of these two plot types. Alternatively we could re-add them here, and make the behavior such that if enumerable-statistics is included it will be faster, otherwise it will still work.